refactor: extract Source Ports into dedicated tab in Settings dialog
Split the Settings dialog's 'Paths' tab into two tabs:
- New 'Source Ports' tab (scan/get/add/edit/delete source ports,
  extracted verbatim into SourcePortsTab.tsx along with its
  previously-inline PortForm sub-component)
- Slimmed 'Paths' tab now holds only the four directory-path fields
  (WAD Assets, Mod Logic, Telemetry/Saves, Optical/Screens)

Pure extraction/relocation, no behavior change. Verified via real
Electron launch that both tabs render and function identically.

Prep work for adding FreeDoom download feature to WAD Config,
which belongs there rather than in the now-overloaded Paths tab.
